add models and controllers folder structure and minor refactoring
